# resources/spec/api-v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UnityGameTranslator API v1",
    "version": "1",
    "summary": "What a Core (the mod, the Manager, a second engine's Core) and the site say to each other over HTTP.",
...
    "x-rules": {
      "absent-means-unknown": "A missing key is 'not said', never a value. Clients keep the previous value; servers may omit any field marked additive.",
      "error-body": "{error?: string, message?: string, reason?: string, refused_code?: string, errors?: {field: [string]}} — show error or message as-is.",
      "languages": "Catalogue names only (catalogs/languages.json), never ISO codes; 'auto' is refused where a language is required.",
      "declarations": "User-Agent, X-UGT-Game and X-UGT-Device are read for display and grouping of the caller's own lines; never for authorisation."
    }
  },
  "servers": [
    { "url": "https://unitygametranslator.asymptomatikgames.com/api/v1", "description": "Production. The mod and the Manager compile this in (Directory.Build.props); config.json's api_base_url overrides it for a self-hosted site." }
  ],
...
  "security": [],
...
  "paths": {
...
    "/me/notifications": {
      "get": {
        "tags": ["account"],
        "operationId": "notifications",
        "summary": "Unread notifications, summarised in English for an overlay.",
        "description": "`unread` counts them all; `items` carries the first five, each with a pre-rendered `text` and a `url` to open, or null.",
        "security": [{ "bearer": [] }],
        "x-throttle": "30/min per token",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NotificationsResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Banned" },
          "429": { "$ref": "#/components/responses/TooMany" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "description": "`Authorization: Bearer ugt_…` — the token the device flow delivered. A missing or unknown token is 401 `{error: \"Unauthorized\", message}`; a banned account is 403 `{error: \"Account banned\", message, reason}` on every route but `DELETE /auth/token`. A client that receives 401 on a call it sent a token with signs itself out; a 403 does not mean that unless `error` is exactly `Account banned`." }
    },
...
    "responses": {
      "Unauthorized": { "description": "No token, or a token the site does not know: `{error: \"Unauthorized\", message: \"API token required\" | \"Invalid API token\"}`. The client signs out.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } },
      "Banned": { "description": "`{error: \"Account banned\", message, reason}` — the one 403 a client reads as 'sign out'.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } },
...
      "TooMany": { "description": "Throttled. `Retry-After` in seconds; the client waits it out rather than reporting a failure.", "headers": { "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorBody" } } } },
...
    },
    "schemas": {
...
      "NotificationType": { "type": "string", "enum": ["branch_submitted", "branch_merged", "branches_closed", "branch_orphaned", "announcement", "unknown"] },
...
      "RefusedCode": { "type": "string", "enum": ["branches_refused", "main_abandoned", "main_gone", "branch_frozen"], "description": "Why an upload was refused with 403, beside the sentence: the Main takes no contributions; the Main's owner erased their account; the Main was removed while branches remain; the caller's own branch is frozen (its Main closed since)." },
...
      "ErrorBody": {
        "type": "object",
        "description": "What a refused call answers. `error` and `message` are sentences to show as they are; the mod shows `error` first, then `message`. `reason` comes with a ban; `refused_code` with an upload refusal; `errors` with a Laravel validation failure.",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "reason": { "type": "string" },
          "refused_code": { "$ref": "#/components/schemas/RefusedCode" },
          "errors": { "type": "object", "additionalProperties": { "type": "array", "items": { "type": "string" } } },
          "details": { "type": "array", "items": { "type": "string" } },
          "hint": { "type": "string" }
        },
        "anyOf": [{ "required": ["error"] }, { "required": ["message"] }]
      },
...
      "NotificationsResponse": {
        "type": "object",
        "required": ["unread", "items"],
        "properties": {
          "unread": { "type": "integer" },
          "items": { "type": "array", "maxItems": 5, "items": {
            "type": "object",
            "required": ["id", "type", "text", "url", "uuid"],
            "properties": {
              "id": { "type": "string", "description": "The notification's uuid." },
              "type": { "$ref": "#/components/schemas/NotificationType" },
              "text": { "type": "string", "description": "Pre-rendered English." },
              "url": { "type": ["string", "null"] },
              "uuid": { "type": ["string", "null"], "description": "The lineage this is about, when it is about one (a contribution sent, closed or orphaned); null otherwise. A client holding that lineage answers it on its own screen rather than on the site." }
            }
          } }
        }
      },
...
    }
  }
}